requirements.txt
python-telegram-bot==21.4
google-generativeai==0.8.3
ffmpeg-python==0.2.0
python-dotenv==1.0.1
faster-whisper==1.0.3
